==> scripts/01-cleaning-data.ipynb <==
# imports

import pandas as pd
import numpy as np
import re

from helper.data_cleaning import find_highest_value, extract_cpv, convert_to_euro

# Reading and cleaning parsed data. 

This notebook reads parsed tender data and cleans the dataset.
It creates the "parsed_data_new.csv" file. 
Saved dataset is further processed in the 02-descriptives notebook.

## Reading data

columns = ["id", "title", "cpv", "contract_type", "description", "country", "contracting_authority_type", "contracting_authority_activity", "value", "type_of_procedure", 
           "government_procurement_agreement", "framework_agreement", "notice_dispatch_date", "application_date", "tender_opening_date"]

# read data
df = pd.read_csv('../data/parsed_data_new.csv', encoding='utf-8', names=columns, low_memory=False)

print(f"There are a total of {len(df)} tenders in the dataset.")

df.head()

# cleaning df

# Value. Since the value column contains a lot of information, we need to extract the currency and the highest value from the column.

european_currencies = "EUR|GBP|CZK|HRK|DKK|PLN|RON|SEK|CHF|ISK|NOK|HUF|BAM|MKD|ALL|UAH|MDL|RSD|BGN"
df['currency'] = df['value'].replace(np.nan, "").apply(lambda x: re.findall(european_currencies, x)[0] if re.findall(european_currencies, x) else None)
df['value_hi'] = df['value'].apply(find_highest_value)

# CPV. Extract all numbers from column cpv and save them in a list.
df["cpv"] = df["cpv"].apply(extract_cpv)

# Year. Extract year from notice_dispatch_date, application_date, tender_opening_date and id.

df["notice_dispatch_date"] = df["notice_dispatch_date"].str.extract('(\d{4})')
df["application_date"] = df["application_date"].str.extract('(\d{4})')
df["tender_opening_date"] = df["tender_opening_date"].str.extract('(\d{4})')
df["id-year"] = df["id"].str.split("-").str[1]
df["year"] = df["notice_dispatch_date"].fillna(df["application_date"]).fillna(df["tender_opening_date"]).fillna(df["id-year"])

# Contracting authority type.

df["contracting_authority_type"] = df["contracting_authority_type"].str.replace('[^\w\s]','').str.replace('\n', '')
df["contracting_authority_type"] = df["contracting_authority_type"].str.replace('Other.*', 'Other', regex=True)
df["contracting_authority_type"] = df["contracting_authority_type"].str.replace('Ministry or any other national or federal authority.*', 
                                                                                'Ministry or any other national or federal authority', regex=True)
df["contracting_authority_type"] = df["contracting_authority_type"].str.replace('European institution.*|European Institution.*', 
                                                                                'European institution/agency or international organisation', regex=True)
df["contracting_authority_type"] = df["contracting_authority_type"].str.replace('agencyoffice', 'agency/office', regex=True)


# Contacting authority activity.

df["contracting_authority_activity"] = df["contracting_authority_activity"].str.replace('Other.*', 'Other', regex=True)
df["contracting_authority_activity"] = df["contracting_authority_activity"].str.replace('[^\w\s]','').str.replace('\n', '')
#split string in contracting_authority_activity column by capital letter and save the first word in a new column
df["contracting_authority_activity"] = df["contracting_authority_activity"].str.split('(?=[A-Z])').str[1]


# Government procurement agreement.

df["government_procurement_agreement"] = df["government_procurement_agreement"].str.replace('.*yes?.|.*Yes?.', 'Yes', regex=True)
df["government_procurement_agreement"] = df["government_procurement_agreement"].str.replace('.*no?.|.*No?.', 'No', regex=True)


# Recoding different types of proceduresin dummy variables - and keeping NaNs.

df["is_framework_agreement"] = np.where(df['framework_agreement'].isna(), None, 
                                        np.where(df['framework_agreement'].str.contains("framework agreement"), "Yes", "No"))


df["is_dynamic"] = np.where(df['framework_agreement'].isna(), None, 
                                        np.where(df['framework_agreement'].str.contains("dynamic purchasing system"), "Yes", "No"))

df["is_open"] = np.where(df['type_of_procedure'].isna(), None, 
                                        np.where(df['type_of_procedure'].str.contains("Open"), "Yes", "No"))


# recode variable "contact_type". Where str.contains "Servic" recode to "Services", where str.contains "Suppli" recode to "Supplies", where str.contains 
# "Works" recode to "Works". If none of the above, recode to np.nan. Save as column contract_type.

df["contract_type"] = np.where(df['contract_type'].str.contains("Servic"), "Services", 
                                 np.where(df['contract_type'].str.contains("Suppli"), "Supplies",
                                            np.where(df['contract_type'].str.contains("Works"), "Works", None)))


# Getting information on buyer.

df['buyer_name'] = df['country'].str.extract(r'Official name: (.*?)<br/>', expand=False)
df['buyer_address'] = df['country'].str.extract(r'Postal address: (.*?)<br/>', expand=False)
df['buyer_town'] = df['country'].str.extract(r'Town: (.*?)<br/>', expand=False)
# Country comes with two types of breaks. We need to extract the country from both types of breaks.
df['country'] = df['country'].str.extract(r'Country: (.*?)<br/>', expand=False)
df["country"] = df["country"].str.split("</p>").str[0]

# getting tender value in EUR.

df["value_eur"] = df.apply(convert_to_euro, axis=1)


# keep only relevant columns: id, year, title, cpv, contract_type, description, country, 
# buyer_name, buyer_address, buyer_town, contracting_authority_type, contracting_authority_activity, 
# is_framework_agreement, is_dynamic, is_open, value_eur

df = df[["id", "year", "title", "cpv", "contract_type", "description", "country", "buyer_name", "buyer_address", "buyer_town",
            "contracting_authority_type", "contracting_authority_activity", "government_procurement_agreement", "is_framework_agreement", "is_dynamic", "is_open", "value_eur"]]

df.head()

# Duplicate check

Note: there will be no duplicates based on the id - these were removed already in the stage of html download.
Based on a manual check, current duplicates in the dataset are based on either a) rows that have all or many missing data b) tenders that were published multiple times as there were no suppliers.

# number of rows with missing values in all of: 
# title, description, contracting_authority_type, contracting_authority_activity, government_procurement_agreement, is_framework_agreement, is_dynamic, is_open, value_eur

df_missing = df[(df["title"].isna()) & (df["description"].isna())]

# show na in columns of df2

df_missing.isna().sum()

print("Number of rows with missing 80% of data: ", len(df[(df["title"].isna()) & (df["description"].isna())]))

# Drop rows with missing data in both title and description.

df = df.dropna(subset=["title", "description"], how="all")

duplicates = df[df.duplicated(subset=["title", "description", "country", "buyer_name", "buyer_address", "buyer_town", "value_eur", "year"])]

print("Number of rows that are duplicates: ", len(duplicates))

duplicates_sum = duplicates.groupby(["title", "description", "country", "buyer_name", "buyer_address", "buyer_town", "value_eur", "year"]).size().sort_values(ascending=False)
duplicates_sum

# droping duplicates

df = df.drop_duplicates(subset=["title", "description", "country", "buyer_name", "buyer_address", "buyer_town", "value_eur", "year"])

# Saving deduplicated data

# save cleaned df

df.to_csv('../data/cleaned_tender_data.csv', index=False)

# Simple value check

df = pd.read_csv('../data/cleaned_tender_data.csv')

print("Total number of records: ", len(df))

# for values in type of procedure, print unique values with their counts, including nan

df["is_open"].value_counts(dropna=False)

df["contract_type"].value_counts(dropna=False)

print(f"There are only {len(df[df['year'] == 2013])} records from 2013.")

df["year"].value_counts(dropna=False)

df["government_procurement_agreement"].value_counts(dropna=False)

df["contracting_authority_type"].value_counts(dropna=False)

# save value counts as csv

df["contracting_authority_activity"].value_counts(dropna=False)

df["is_framework_agreement"].value_counts(dropna=False)

df["is_dynamic"].value_counts(dropna=False)

print(f"There are {len(df['country'].value_counts()[df['country'].value_counts() < 500])} countries with less than 500 records.")

print(df["country"].value_counts(dropna=False).to_string())
