==> scripts/03-missing-data-analysis.ipynb <==
import pandas as pd
import numpy as np
from helper.missing_analysis import missing_values_table

# Missing data analysis.

This notebook explores missing data at the tender level.

# import cleaned data
df = pd.read_csv('../data/cleaned_recoded_tender_data.csv')
df.head()

# save sum missing values of contracting authority type and activity as a variable

sum_missing_contracting_authority_type_pre = df["contracting_authority_type"].isnull().sum()
sum_missing_contracting_authority_activity_pre = df["contracting_authority_activity"].isnull().sum()


# create a nested dictionary called buyers from df with buyer_name as main key, and address = buyer_address, 
# town = buyer_town, type = contracting_authority_type as value, activity = contracting_authority_activity as subkeys.

print("Creating a dictionary of contracting authorities...")

buyers = {}
for index, row in df.iterrows():
    if row["buyer_name"] not in buyers:
        buyers[row["buyer_name"]] = {"address": row["buyer_address"], "town": row["buyer_town"], "type": row["contracting_authority_type"], "activity": row["contracting_authority_activity"]}
    else:
        continue

# function to replace missing data for contracting authority.
# if contracting authority type is missing, check if buyer_name is in buyers dictionary. If it is, copy the
# value of "type" from the dictionary to the df. if contracting authority activity is missing, check if buyer_name is in buyers dictionary. If it is, copy the
# value of "activity" from the dictionary to the df.

def fill_missing_contracting_authority_values(row):

    if pd.isnull(row["contracting_authority_type"]):
        if row["buyer_name"] in buyers:
            row["contracting_authority_type"] = buyers[row["buyer_name"]]["type"]
        else:
            pass
    else:
        pass
    
    if pd.isnull(row["contracting_authority_activity"]):
        if row["buyer_name"] in buyers:
            row["contracting_authority_activity"] = buyers[row["buyer_name"]]["activity"]
        else:
            pass
    else:
        pass
    
    return row

print("Filling missing values in contracting authority type and activity...\n")

# apply function to df
df = df.apply(fill_missing_contracting_authority_values, axis=1)

sum_missing_contracting_authority_type_post = df["contracting_authority_type"].isnull().sum()
sum_missing_contracting_authority_activity_post = df["contracting_authority_activity"].isnull().sum()

# print difference in number of missing values before and after function

print("Values filled for contracting authority type: ", sum_missing_contracting_authority_type_pre - sum_missing_contracting_authority_type_post)
print("Values filled for contracting authority activity: ", sum_missing_contracting_authority_activity_pre - sum_missing_contracting_authority_activity_post)



missing = missing_values_table(df)
missing

# print number of rows that have no missing in any of these columns: value_eur, is_open, government_procurement_agreement
# is_framework_agreement, is_dynamic, contracting_authority_type, contracting_authority_activity, 
# contract_type

print("Number of rows with no missing values in any of key columns: ", df[df["value_eur"].notnull() & df["is_open"].notnull() & df["government_procurement_agreement"].notnull() & df["is_framework_agreement"].notnull() & df["is_dynamic"].notnull() & df["contracting_authority_type"].notnull() & df["contracting_authority_activity"].notnull() & df["contract_type"].notnull()].shape[0])
print("Number of rows with no missing values in any of key columns, but tender value: ", df[df["is_open"].notnull() & df["government_procurement_agreement"].notnull() & df["is_framework_agreement"].notnull() & df["is_dynamic"].notnull() & df["contracting_authority_type"].notnull() & df["contracting_authority_activity"].notnull() & df["contract_type"].notnull()].shape[0])
print("Total number of rows: ", len(df))

# Missing data mechanism

Testing correlations between variables and missingness.

# create a new dataframe where for every column with missing values there is a variable with _missing extension,
# where value is 1 is value is missing, 0 if not.

df_missing = df.copy()

# removing unnecessary values

df_missing = df_missing[df_missing["country"].isin(df_missing["country"].value_counts()[df_missing["country"].value_counts() > 500].index)]
#remove year 2013
df_missing = df_missing[df_missing["year"] != 2013]

# from df_missing, drop id, title, cpv, contract_type, description, buyer_name, buyer_address, buyer_town,
# contracting_authority_type, contracting_authority_activity, is_open, is_dynamic, is_framework_agreement

for col in df_missing.columns:
    if df_missing[col].isnull().sum() > 0 and col not in ["title", "cpv", "description", "buyer_name", "buyer_address", "buyer_town"]:
        df_missing[col + "_missing"] = df_missing[col].isnull().astype(int)
    else:
        continue

df_missing = df_missing.drop(["id", "title", "cpv", "description", "buyer_name", "buyer_address", "buyer_town"], axis=1)


# create dummy variables from year, country, saas

df_missing = pd.get_dummies(df_missing, columns=["year", "country", "saas"])


### ADDED FOR SUPER LONG
##df_missing = pd.get_dummies(df_missing, columns=["contract_type", "contracting_authority_type", "contracting_authority_activity", "is_open", "is_dynamic", "is_framework_agreement", "government_procurement_agreement"])

# correlate all variables in df_missing that end with _missing with
# all columns that start with year, country, saas
# save in dataframe called "correlation_matrix"
# save correlation coefficients in column with subscript _r, p value in column with subscript _p

from scipy.stats import pearsonr

correlation_matrix = pd.DataFrame(columns=["contract_type_missing_r"])
for col in df_missing.columns:
    if col.endswith("_missing"):
        for col2 in df_missing.columns:
            if col2.startswith("year") or col2.startswith("country") or col2.startswith("saas"):
                
                p_value = pearsonr(df_missing[col2], df_missing[col])[1]         
               
                correlation_matrix.loc[col2, col + "_r"] = round(pearsonr(df_missing[col2], df_missing[col])[0], 2)
            else:
                continue
    else:
        continue

# make column in correlation_matrix with p value less than 0.05 bold

correlation_matrix


correlation_matrix.to_csv("../outputs/correlation_matrix_wmissing.csv")
