==> scripts/05-merging-data.ipynb <==
import pandas as pd
import numpy as np

# Tender and country data merging.

This notebook merges tender and country data.
It creates "cleaned_recoded_tender_country_data.csv" dataset.

# import cleaned data
tender_data = pd.read_csv('../data/cleaned_recoded_tender_data.csv')
tender_data.head()

country_data = pd.read_csv('../data/country_data_all_new.csv')

country_data.head()

# Merge the two dataframes
# Merge on country and year
# From country_data keep column called value, but rename it to the name of the indicator

df = tender_data.merge(country_data, how='left', left_on=['country', 'year'], right_on=['country', 'year'])
df = df.drop(df.columns[0], axis=1)
df.head()

df.columns

#shape of df_analysis

df.shape


# missing data in population column

df['population'].isnull().sum()

df.to_csv('../data/cleaned_recoded_tender_country_data.csv', index=False)
